+++ packages/target/lib/src/modelable.dart
@@ -1,4 +1,4 @@
-import 'package:dartz/dartz.dart';
+import 'package:target/src/either.dart';
 
 abstract interface class Modelable<Failure, Model> {
   /// Creates a model from this.