+++ packages/target/lib/src/raise.dart
@@ -1,16 +1,16 @@
-import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
+import 'package:target/src/either.dart';
 import 'package:target/src/raise_fold.dart';
 
 /// Implementation of arrow-kt raise dsl.
-abstract interface class Raise<Error> {
-  Never raise(Error r);
+abstract interface class Raise<E> {
+  Never raise(E r);
 }
 
-final class RaiseCancellationException<Error> extends Equatable
+final class RaiseCancellationException<E> extends Equatable
     implements Exception {
-  final Error raised;
-  final Raise<Error> raise;
+  final E raised;
+  final Raise<E> raise;
 
   const RaiseCancellationException({required this.raised, required this.raise});
 
@@ -20,45 +20,30 @@ final class RaiseCancellationException<Error> extends Equatable
 
 final class RaiseLeakedError extends StateError {
   RaiseLeakedError()
-      : super(
-          'raise or bind was called outside of its DSL scope, and the DSL Scoped operator was leaked',
-        );
+    : super(
+        'raise or bind was called outside of its DSL scope, and the DSL Scoped operator was leaked',
+      );
 }
 
-A recover<Error, A>(
-  A Function(Raise<Error>) block,
-  A Function(Error) recover,
-) =>
-    fold(block, (it) => throw it, recover, id);
+A recover<E, A>(A Function(Raise<E>) block, A Function(E) recover) =>
+    fold(block, (it) => throw it, recover, (it) => it);
 
-Future<A> recoverAsync<Error, A>(
-  Future<A> Function(Raise<Error>) block,
-  A Function(Error) recover,
-) =>
-    foldAsync(block, (it) => throw it, recover, id);
+Future<A> recoverAsync<E, A>(
+  Future<A> Function(Raise<E>) block,
+  A Function(E) recover,
+) => foldAsync(block, (it) => throw it, recover, (it) => it);
 
-extension RaiseEnsureExtension<Error> on Raise<Error> {
-  A bind<A, E extends Error>(Either<E, A> r) => r.fold(raise, id);
+extension RaiseEnsureExtension<E> on Raise<E> {
+  A bind<A, E2 extends E>(Either<E2, A> r) => r.fold(raise, (it) => it);
 
-  void ensure(bool condition, Error Function() raise) {
+  void ensure(bool condition, E Function() raise) {
     if (!condition) {
       this.raise(raise());
     }
   }
-
-  B ensureNotNull<B>(B? value, Error Function() raise) =>
-      value ?? this.raise(raise());
 }
 
-A merge<A>(A Function(Raise<A>) block) => recover(block, id);
+A merge<A>(A Function(Raise<A>) block) => recover<A, A>(block, (it) => it);
 
 Future<A> mergeAsync<A>(Future<A> Function(Raise<A>) block) =>
-    recoverAsync(block, id);
-
-extension RaiseEitherExtension<Error, Error2 extends Error, A>
-    on Either<Error2, A> {
-  /// Binds this to the given [Raise].
-  ///
-  /// Note: [bind] member function already exists on [Either].
-  A bindTo(Raise<Error> r) => r.bind(this);
-}
+    recoverAsync<A, A>(block, (it) => it);