+++ packages/target/lib/src/raise_builders.dart
@@ -1,17 +1,16 @@
-import 'package:dartz/dartz.dart';
+import 'package:target/src/either.dart';
+import 'package:target/src/option.dart';
 import 'package:target/src/raise.dart';
 import 'package:target/src/raise_fold.dart';
 
 /// DSL for computing an [Either] from the given effect [block]. Based on the
 /// arrow-kt implementation.
-Either<Error, A> either<Error, A>(A Function(Raise<Error> r) block) =>
+Either<E, A> either<E, A>(A Function(Raise<E> r) block) =>
     foldOrThrow(block, Left.new, Right.new);
 
 /// DSL for computing an [Either] async from the given effect [block]. Based on
 /// the arrow-kt implementation.
-Future<Either<Error, A>> eitherAsync<Error, A>(
-  Future<A> Function(Raise<Error> r) block,
-) =>
+Future<Either<E, A>> eitherAsync<E, A>(Future<A> Function(Raise<E> r) block) =>
     foldOrThrowAsync(block, Left.new, Right.new);
 
 /// DSL for computing a nullable from the given effect [block]. Based on the
@@ -32,15 +31,13 @@ final class NullableRaise implements Raise<void> {
   @override
   Never raise(void r) => _delegate.raise(r);
 
-  A bind<B, A>(Either<B, A> r) => r.fold((_) => raise(null), id);
+  A bind<B, A>(Either<B, A> r) => r.fold((_) => raise(null), (it) => it);
 
-  A bindOption<A>(Option<A> r) => r.fold(() => raise(null), id);
+  A bindOption<A>(Option<A> r) => r.fold(() => raise(null), (it) => it);
 
   void ensure(bool condition) {
     if (!condition) {
       raise(null);
     }
   }
-
-  B ensureNotNull<B>(B? value) => value ?? raise(null);
 }