+++ packages/target/test/test_value_validator.dart
@@ -1,9 +1,12 @@
-import 'package:dartz/dartz.dart';
 import 'package:target/target.dart';
 
 /// Test value validator returning the given [result].
-class TestValueValidator<I extends Object, F extends ValueFailure<I>,
-    T extends ValueObject<I>> extends ValueValidator<I, F, T> {
+class TestValueValidator<
+  I extends Object,
+  F extends ValueFailure<I>,
+  T extends ValueObject<I>
+>
+    extends ValueValidator<I, F, T> {
   final Either<F, T> result;
 
   const TestValueValidator(this.result);