+++ packages/target/test/value_validator_test.dart
@@ -1,17 +1,19 @@
-import 'package:dartz/dartz.dart';
+import 'package:target/src/either.dart';
 import 'package:target/src/generic_value_object.dart';
+import 'package:target/src/option.dart';
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
 
 import 'test_value_validator.dart';
 
 void main() {
-  const testResult = Right<Never, _TestValueObject>(_TestValueObject());
+  const testResult = Right<_TestValueObject>(_TestValueObject());
 
   group('ValueValidator.nullable', () {
     test('returns Right(null) when input is null', () {
-      const valueValidator =
-          TestValueValidator<int, Never, _TestValueObject>(testResult);
+      const valueValidator = TestValueValidator<int, Never, _TestValueObject>(
+        testResult,
+      );
 
       final result = valueValidator.nullable(null);
 
@@ -20,8 +22,9 @@ void main() {
   });
   group('ValueValidator.option', () {
     test('returns Right(None()) when the input is None()', () {
-      const valueValidator =
-          TestValueValidator<int, Never, _TestValueObject>(testResult);
+      const valueValidator = TestValueValidator<int, Never, _TestValueObject>(
+        testResult,
+      );
 
       final result = valueValidator.option(const None());
 
@@ -30,16 +33,18 @@ void main() {
   });
   group('ValueValidator.nullableOption', () {
     test('returns Right(None()) when the input is None()', () {
-      const valueValidator =
-          TestValueValidator<int, Never, _TestValueObject>(testResult);
+      const valueValidator = TestValueValidator<int, Never, _TestValueObject>(
+        testResult,
+      );
 
       final result = valueValidator.nullableOption(const None());
 
       expect(result, const Right(None()));
     });
     test('returns Right(Some(null)) when the input is Some(null)', () {
-      const valueValidator =
-          TestValueValidator<int, Never, _TestValueObject>(testResult);
+      const valueValidator = TestValueValidator<int, Never, _TestValueObject>(
+        testResult,
+      );
 
       final result = valueValidator.nullableOption(const Some(null));