+++ packages/target_annotation_processor/test/source_gen_src/test_nested_model.dart
@@ -2,8 +2,7 @@ import 'package:source_gen_test/source_gen_test.dart';
 import 'package:target/target.dart';
 import 'package:target_annotation/target_annotation.dart';
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 // Generated code. Do not modify by hand.
 
 // ignore_for_file: require_trailing_commas, unused_element
@@ -15,20 +14,20 @@ Either<Nel<ModelFieldFailure>, Model> _$of({
 }) {
   final vId = PositiveInt.of(id);
   final vField = PositiveInt.of(field);
-  if (vId is Right<GenericValueFailure<int>, PositiveInt> &&
-      vField is Right<GenericValueFailure<int>, PositiveInt> &&
-      parent is Right<Nel<ModelFieldFailure>, Model>) {
+  if (vId is Right<PositiveInt> &&
+      vField is Right<PositiveInt> &&
+      parent is Right<Model>) {
     return Right(
       Model(id: vId.value, field: vField.value, parent: parent.value),
     );
   } else {
     return Left(
       Nel.fromListUnsafe([
-        if (vId is Left<GenericValueFailure<int>, PositiveInt>)
+        if (vId is Left<GenericValueFailure<int>>)
           ModelFieldFailureId(vId.value),
-        if (vField is Left<GenericValueFailure<int>, PositiveInt>)
+        if (vField is Left<GenericValueFailure<int>>)
           ModelFieldFailureField(vField.value),
-        if (parent is Left<Nel<ModelFieldFailure>, Model>)
+        if (parent is Left<Nel<ModelFieldFailure>>)
           ModelFieldFailureParent(parent.value),
       ]),
     );
@@ -58,8 +57,7 @@ final class ModelFieldFailureParent
     extends ModelFieldFailure<Nel<ModelFieldFailure>> {
   const ModelFieldFailureParent(super.parent);
 }
-''',
-)
+''')
 @validatable
 final class Model {
   final PositiveInt id;