+++ packages/target_extension/lib/src/target_iterable_extensions.dart
@@ -1,4 +1,3 @@
-import 'package:dartz/dartz.dart';
 import 'package:target_extension/src/build_map.dart';
 
 extension TargetIterableExtensions<T> on Iterable<T> {
@@ -14,19 +13,5 @@ extension TargetIterableExtensions<T> on Iterable<T> {
         );
       }
     });
-  }
-
-  /// Traverse either implementation based on the arrow-kt library in Kotlin.
-  Either<L, List<R>> traverseEither<L, R>(Either<L, R> Function(T it) f) {
-    final destination = <R>[];
-    for (final item in this) {
-      final res = f(item);
-      if (res is Right<L, R>) {
-        destination.add(res.value);
-      } else if (res is Left<L, R>) {
-        return Left(res.value);
-      }
-    }
-    return Right(destination);
   }
 }