+++ packages/target_extension/lib/src/target_option_extensions.dart
@@ -1,14 +1,10 @@
-import 'package:dartz/dartz.dart';
-import 'package:target_extension/src/target_effects.dart';
+import 'package:target/target.dart';
 
 extension TargetOptionX<T> on Option<T> {
-  /// Performs the given side-[effect] if this is [Some].
-  void ifSome(DartzEffect<T> effect) {
-    if (this is Some<T>) {
-      effect((this as Some<T>).value);
-    }
-  }
+  Either<L, T> toEither<L>(L Function() onNone) =>
+      this.fold(() => Left(onNone()), Right.new);
 
   /// Returns `Some(null)` when [None].
-  Some<T?> orSomeNull() => fold(() => const Some(null), (_) => this as Some<T>);
+  Some<T?> orSomeNull() =>
+      this.fold(() => const Some(null), (_) => this as Some<T>);
 }